refactor: reuse retry logic
